requirements.txt
Django>=2.0
graphene_django
django-filter>=2
django-cors-headers
gunicorn
django-heroku
django-graphql-jwt
django-extensions